rename chat_with_pdf to workbench
